Admin validation added

+++ app/api/bankAccount/(admin)/[id]/route.ts
@@ -3,6 +3,8 @@ import dbConnect from "@/app/db/connection";
 import { BankAccount } from "@/app/models/BankAccount";
 import User from "@/app/models/User";
 import AccountType from "@/app/models/AccountType";
+import { getServerSession } from "next-auth";
+import { authOptions } from "@/app/api/auth/[...nextauth]/route";
 
 dbConnect();
 
@@ -16,6 +18,14 @@ export async function GET(request: NextRequest, params: params) {
   const id = params.params.id;
 
   try {
+    const session = await getServerSession(authOptions);
+    // Verify if the user is authenticated and is an admin
+    if (!session?.user || session.user.role !== "admin") {
+      return new NextResponse("Unauthorized", {
+        status: 401,
+      });
+    }
+
     const users = await User.find();
     const accountType = await AccountType.find();
 
@@ -23,11 +33,7 @@ export async function GET(request: NextRequest, params: params) {
       .populate("client", "name username email dpi address phone work")
       .populate("accountType", "name description");
 
-    const data = {
-      bankAccount,
-    };
-
-    // Validar si no se encontró la notificación
+    // Validate if the bank account was not found
     if (!bankAccount) {
       return new NextResponse("Account not found", {
         status: 404,
@@ -50,11 +56,19 @@ export async function PUT(request: Request, params: params) {
   const data = await request.json();
 
   try {
+    const session = await getServerSession(authOptions);
+    // Verify if the user is authenticated and is an admin
+    if (!session?.user || session.user.role !== "admin") {
+      return new NextResponse("Unauthorized", {
+        status: 401,
+      });
+    }
+
     const bankAccount = await BankAccount.findByIdAndUpdate(id, data, {
       new: true,
     });
 
-    // Validar si no se encontró la notificación
+    // Validate if the bank account was not found
     if (!bankAccount) {
       return new NextResponse("Account not found", {
         status: 404,
@@ -77,9 +91,17 @@ export async function DELETE(request: Request, params: params) {
   const id = params.params.id;
 
   try {
+    const session = await getServerSession(authOptions);
+    // Verify if the user is authenticated and is an admin
+    if (!session?.user || session.user.role !== "admin") {
+      return new NextResponse("Unauthorized", {
+        status: 401,
+      });
+    }
+
     const bankAccount = await BankAccount.findByIdAndDelete(id);
 
-    // Validar si no se encontró la notificación
+    // Validate if the bank account was not found
     if (!bankAccount) {
       return new NextResponse("Account not found", {
         status: 404,
